Implemented some changes to hashing and create and edit user to hopefully increase performance on these routes since these were very slow.

See merge request tddc88-2022/c4/rdx-solutions-backend-project!104

+++ Web/backend/views/userView.py
@@ -1,20 +1,22 @@
 from backend.__init__ import serviceInjector as si
 from backend.services.userService import UserService
+from ..serializers import UserInfoSerializer
 
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from knox.auth import TokenAuthentication
+from rest_framework.permissions import IsAdminUser
 
-from django.http import Http404, HttpResponseBadRequest
-from ..serializers import UserInfoSerializer
+from django.core.exceptions import PermissionDenied
+from django.http import Http404
 
 
 class User(APIView):
     '''View for getting all users and posting a user to the DB. 
     Should only be accessable by admins'''
 
-#    authentication_classes = (TokenAuthentication,)
+    permission_classes = [IsAdminUser]
+
     @si.inject
     def __init__(self, _deps, *args):
         self.userService: UserService = _deps['UserService']()
@@ -34,22 +36,18 @@ class User(APIView):
     def post(self, request):
         '''posts a user to the database. Should Only accessable by Admins'''
 
-        json_body = request.data
-        try:
-            barcode_id = json_body['barcodeId']
-            nfcId = json_body['nfcId']
-            username = json_body['username']
-            password = json_body['password']
-            cost_centers = json_body['costCenters']
-            group = json_body['role']
-        except:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
+        barcode_id = request.data.get('barcodeId', None)
+        nfc_id = request.data.get('nfcId', None)
+        username = request.data.get('username')
+        password = request.data.get('password')
+        cost_center = request.data.get('costCenters')
+        group = request.data.get('role')
 
         new_user = self.userService.create_user(
-            username, password, barcode_id, nfcId, cost_centers, group)
+            username, password, barcode_id, nfc_id, cost_center, group)
 
         if new_user is None:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
+            return Response('user could not be created', status=status.HTTP_400_BAD_REQUEST)
 
         serializer = UserInfoSerializer(new_user)
         if serializer.is_valid:
@@ -58,9 +56,10 @@ class User(APIView):
 
 
 class UserId(APIView):
-    '''User'''
+    '''User view. Can only be accessed by admins'''
 
- #   authentication_classes = (TokenAuthentication,)
+    permission_classes = [IsAdminUser]
+
     @si.inject
     def __init__(self, _deps, *args):
         self.userService: UserService = _deps['UserService']()
@@ -68,9 +67,9 @@ class UserId(APIView):
     def get(self, request, user_id):
         '''Get.'''
         if request.method == 'GET':
-            # A user can see a group if they have permission
-            #            if not request.user.has_perm('backend.view_group'):
-         #               raise PermissionDenied
+            if not request.user.has_perm('backend.view_group'):
+                raise PermissionDenied
+
             user = self.userService.get_user_info(user_id)
             if user is None:
                 raise Http404("Could not find user")
@@ -81,34 +80,23 @@ class UserId(APIView):
 
     def put(self, request, user_id):
         '''Edit user'''
-        json_body = request.data
-        try:
-            barcode_id = json_body['barcodeId']
-        except:
-            barcode_id = None
-        try:
-            nfc_id = json_body['nfcId']
-        except:
-            nfc_id = None
-        try:
-            username = json_body['username']
-        except:
-            username = None
-        try:
-            password = json_body['password']
-        except:
-            password = None
-        try:
-            cost_center = json_body['costCenters']
-        except:
-            cost_center = None
-        try:
-            group = json_body['role']
-        except:
-            group = None
+        user = self.userService.get_user_info(user_id)
+        if user is None:
+            return Response({'User does not exist'}, status=status.HTTP_400_BAD_REQUEST)
+
+        barcode_id = request.data.get('barcodeId', user.barcode_id)
+        nfc_id = request.data.get('nfcId', user.nfc_id)
+        username = request.data.get('username', user.user.username)
+        password = request.data.get('password', user.user.password)
+        cost_center = request.data.get('costCenters', user.cost_center)
+        cost_center = request.data.get('costCenters', user.cost_center)
+        group = request.data.get('role', user.group)
 
         updated_user = self.userService.update_user(user_id=user_id, barcode_id=barcode_id, nfc_id=nfc_id, username=username,
                                                     password=password, cost_center=cost_center, group=group)
+
+        if updated_user is None:
+            return Response({'user could not be updated'}, status=status.HTTP_400_BAD_REQUEST)
 
         serialized_order = UserInfoSerializer(updated_user)
         if serialized_order.is_valid:
@@ -117,6 +105,10 @@ class UserId(APIView):
 
     def delete(self, request, user_id):
         '''Delete order func'''
+        user = self.userService.get_user_info(user_id)
+        if user is None:
+            return Response({'Deletion failed. User does not exist'}, status=status.HTTP_400_BAD_REQUEST)
+
         deleted = self.userService.delete_user(user_id=user_id)
         if deleted is None:
             return Response({'User deletion failed'}, status=status.HTTP_400_BAD_REQUEST)
